hardware/requirements.txt
RPi.GPIO==0.7.1
pymongo==4.10.1
